Integrated run and save to one file
